requirements.txt
pymupdf
openai
python-dotenv
tqdm